Labels, sheathing panels amendments
--- amendments courtesy of Paul Boden, University of Wolverhampton  -----

Part SKI/A – An additional double width slot has been included to allow them to connect to the R1,R2, F2 & F6 parts
CEIL/1 – One of the slot was undersize in width, so copied and mirrored this slot to match the others. The discrepancy prevented the Ceiling panels to assemble.

Additional amendment included on Sheet 3 – Part FRA-A/S4/EX, should now read FRA-A/S4/IN to match the assembly instructions.