Edited DXF and SKP files of the chassis
On the DXF: modified slots on the Ski/B to avoid interference with the
reinforcers FRA-A/R2. Shortened an Ext. Wall part of the edge of the
roof and copied it to cut the needed amount.

On the SKP:  Modified slots on Ski/A and Ski/B. Added missing part on
one edge of external sheating, shortened same parts of the other side to
avoid interference and added missing copies to the arrangement of parts
to be cut. Align parts of End Walls.